+++ lib/src/dart_date.dart
@@ -21,19 +21,23 @@ class Interval {
   DateTime get end => _start.add(_duration);
 
   bool includes(DateTime date) =>
-      ((date.isAfter(start) || date.isAtSameMomentAs(start)) && (date.isBefore(end) || date.isAtSameMomentAs(end)));
+      ((date.isAfter(start) || date.isAtSameMomentAs(start)) &&
+          (date.isBefore(end) || date.isAtSameMomentAs(end)));
 
-  bool contains(Interval interval) => (includes(interval.start) && includes(interval.end));
+  bool contains(Interval interval) =>
+      (includes(interval.start) && includes(interval.end));
 
   bool cross(Interval other) => (includes(other.start) || includes(other.end));
 
-  bool equals(Interval other) => (start.isAtSameMomentAs(other.start) && end.isAtSameMomentAs(other.end));
+  bool equals(Interval other) =>
+      (start.isAtSameMomentAs(other.start) && end.isAtSameMomentAs(other.end));
 
   Interval union(Interval other) {
     if (cross(other)) {
       if (end.isAfter(other.start) || end.isAtSameMomentAs(other.start)) {
         return Interval(start, other.end);
-      } else if (other.end.isAfter(start) || other.end.isAtSameMomentAs(start)) {
+      } else if (other.end.isAfter(start) ||
+          other.end.isAtSameMomentAs(start)) {
         return Interval(other.start, end);
       } else {
         throw RangeError('Error this: $this; other: $other');
@@ -47,7 +51,8 @@ class Interval {
     if (cross(other)) {
       if (end.isAfter(other.start) || end.isAtSameMomentAs(other.start)) {
         return Interval(other.start, end);
-      } else if (other.end.isAfter(start) || other.end.isAtSameMomentAs(start)) {
+      } else if (other.end.isAfter(start) ||
+          other.end.isAtSameMomentAs(start)) {
         return Interval(other.start, end);
       } else {
         throw RangeError('Error this: $this; other: $other');
@@ -97,11 +102,13 @@ class Interval {
   // Operators
   bool operator <(Interval other) => (start.isBefore(other.start));
 
-  bool operator <=(Interval other) => (start.isBefore(other.start) || start.isAtSameMomentAs(other.start));
+  bool operator <=(Interval other) =>
+      (start.isBefore(other.start) || start.isAtSameMomentAs(other.start));
 
   bool operator >(Interval other) => (end.isAfter(other.end));
 
-  bool operator >=(Interval other) => (end.isAfter(other.end) || end.isAtSameMomentAs(other.end));
+  bool operator >=(Interval other) =>
+      (end.isAfter(other.end) || end.isAtSameMomentAs(other.end));
 
   @override
   String toString() => '<${start} | ${end} | ${duration} >';
@@ -133,7 +140,9 @@ extension Date on DateTime {
     bool isUTC = false,
   }) {
     initializeDateFormatting();
-    return pattern == null ? DateTime.parse(dateString) : DateFormat(pattern, locale).parse(dateString, isUTC);
+    return pattern == null
+        ? DateTime.parse(dateString)
+        : DateFormat(pattern, locale).parse(dateString, isUTC);
   }
 
   /// Create a [Date] object from a Unix timestamp
@@ -261,10 +270,12 @@ extension Date on DateTime {
   int compare(DateTime other) => compareTo(other);
 
   /// Returns true if left [isBefore] than right
-  static DateTime min(DateTime left, DateTime right) => (left < right) ? left : right;
+  static DateTime min(DateTime left, DateTime right) =>
+      (left < right) ? left : right;
 
   /// Returns true if left [isAfter] than right
-  static DateTime max(DateTime left, DateTime right) => (left < right) ? right : left;
+  static DateTime max(DateTime left, DateTime right) =>
+      (left < right) ? right : left;
 
   /// Compare the two dates and return 1 if the first date [isAfter] the second,
   /// -1 if the first date [isBefore] the second or 0 first date [isEqual] the second.
@@ -280,7 +291,8 @@ extension Date on DateTime {
 
   /// Compare the two dates and return -1 if the first date [isAfter] the second,
   /// 1 if the first date [isBefore] the second or 0 first date [isEqual] the second.
-  static int compareDesc(DateTime dateLeft, DateTime dateRight) => (-1) * compareAsc(dateLeft, dateRight);
+  static int compareDesc(DateTime dateLeft, DateTime dateRight) =>
+      (-1) * compareAsc(dateLeft, dateRight);
 
   // int differenceInCalendarDays(dateLeft, dateRight)
   // int differenceInCalendarISOWeeks(dateLeft, dateRight)
@@ -323,7 +335,8 @@ extension Date on DateTime {
   /// - If [allowFromNow] is passed, format will use the From prefix, ie. a date
   ///   5 minutes from now in 'en' locale will display as '5 minutes from now'
   /// If locales was not loaded previously en would be used use timeago.setLocaleMessages to set them
-  String timeago({String locale, DateTime clock, bool allowFromNow}) => timeago_lib.format(
+  String timeago({String locale, DateTime clock, bool allowFromNow}) =>
+      timeago_lib.format(
         this,
         locale: locale,
         clock: clock,
@@ -514,25 +527,25 @@ extension Date on DateTime {
   bool isEqual(other) => equals(other);
 
   /// Return true if this date day is monday
-  bool get isMonday => day == DateTime.monday;
+  bool get isMonday => weekday == DateTime.monday;
 
   /// Return true if this date day is tuesday
-  bool get isTuesday => day == DateTime.tuesday;
+  bool get isTuesday => weekday == DateTime.tuesday;
 
   /// Return true if this date day is wednesday
-  bool get isWednesday => day == DateTime.wednesday;
+  bool get isWednesday => weekday == DateTime.wednesday;
 
   /// Return true if this date day is thursday
-  bool get isThursday => day == DateTime.thursday;
+  bool get isThursday => weekday == DateTime.thursday;
 
   /// Return true if this date day is friday
-  bool get isFriday => day == DateTime.friday;
+  bool get isFriday => weekday == DateTime.friday;
 
   /// Return true if this date day is saturday
-  bool get isSaturday => day == DateTime.saturday;
+  bool get isSaturday => weekday == DateTime.saturday;
 
   /// Return true if this date day is sunday
-  bool get isSunday => day == DateTime.sunday;
+  bool get isSunday => weekday == DateTime.sunday;
 
   /// Is the given date the first day of a month?
   bool get isFirstDayOfMonth => isSameDay(startOfMonth);
@@ -541,10 +554,12 @@ extension Date on DateTime {
   bool get isFuture => isAfter(DateTime.now());
 
   /// Is the given date the last day of a month?
-  bool get isLastDayOfMonth => isSameDay(nextMonth.startOfMonth.subHours(12).startOfDay);
+  bool get isLastDayOfMonth =>
+      isSameDay(nextMonth.startOfMonth.subHours(12).startOfDay);
 
   /// Is the given date in the leap year?
-  bool get isLeapYear => ((year % 400 == 0) || (year % 4 == 0 && year % 100 != 0));
+  bool get isLeapYear =>
+      ((year % 400 == 0) || (year % 4 == 0 && year % 100 != 0));
 
   /// Return true if this date [isBefore] [Date.now]
   bool get isPast => isBefore(DateTime.now());
@@ -567,7 +582,8 @@ extension Date on DateTime {
   // bool isSameQuarter(dateLeft, dateRight)
 
   /// Check if this date is in the same second than other
-  bool isSameSecond(DateTime other) => secondsSinceEpoch == other.secondsSinceEpoch;
+  bool isSameSecond(DateTime other) =>
+      secondsSinceEpoch == other.secondsSinceEpoch;
 
   // bool isSameWeek(dateLeft, dateRight, [options])
 
@@ -614,7 +630,8 @@ extension Date on DateTime {
   bool get isWeekend => (day == DateTime.saturday || day == DateTime.sunday);
 
   /// Checks if a [DateTime] is within a Rage (two dates that makes an [Interval])
-  bool isWithinRange(DateTime startDate, DateTime endDate) => Interval(startDate, endDate).includes(this);
+  bool isWithinRange(DateTime startDate, DateTime endDate) =>
+      Interval(startDate, endDate).includes(this);
 
   /// Checks if a [DateTime] is within an [Interval]
   bool isWithinInterval(Interval interval) => interval.includes(this);
@@ -847,13 +864,15 @@ extension Date on DateTime {
   DateTime get startOfWeek => subtract(Duration(days: weekday)).startOfDay;
 
   /// Get a [DateTime] representing start of year of this [DateTime] in local time.
-  DateTime get startOfYear => clone.setMonth(DateTime.january, 1, 0, 0, 0, 0, 0);
+  DateTime get startOfYear =>
+      clone.setMonth(DateTime.january, 1, 0, 0, 0, 0, 0);
 
   /// Get the start of a local week-numbering year
   DateTime get startOfWeekYear => startOfYear.startOfWeek;
 
   /// Get the start of a local week-numbering year
-  DateTime get startOfISOWeekYear => DateTime(year, DateTime.january, 4).startOfISOWeek;
+  DateTime get startOfISOWeekYear =>
+      DateTime(year, DateTime.january, 4).startOfISOWeek;
 
   /// Get the number of weeks in an ISO week-numbering year
   int get getISOWeeksInYear {
@@ -901,11 +920,13 @@ extension Date on DateTime {
 
   bool operator <(DateTime other) => isBefore(other);
 
-  bool operator <=(DateTime other) => (isBefore(other) || isAtSameMomentAs(other));
+  bool operator <=(DateTime other) =>
+      (isBefore(other) || isAtSameMomentAs(other));
 
   bool operator >(DateTime other) => isAfter(other);
 
-  bool operator >=(DateTime other) => (isAfter(other) || isAtSameMomentAs(other));
+  bool operator >=(DateTime other) =>
+      (isAfter(other) || isAtSameMomentAs(other));
 
   String toHumanString() => format('E MMM d y H:m:s');